1.3 mvvm 구조로 리팩토링
