+++ .gitignore
@@ -1,0 +1,1 @@
+meme_android/.idea/deploymentTargetDropDown.xml
