+++ meme_android/app/src/main/java/com/likewhile/meme/data/model/MemoItem.kt
@@ -1,8 +1,8 @@
-package com.likewhile.meme
+package com.likewhile.meme.data.model
 
 data class MemoItem(
     var id: Long = -1L,
-    var title: String = "com.likewhile.com.likewhile.meme",
+    var title: String = "meme",
     var content: String,
     var date: String,
     var isFixed: Boolean
