+++ meme_android/app/src/main/java/com/likewhile/meme/data/repository/MemoRepository.kt
@@ -1,0 +1,4 @@
+package com.likewhile.meme.data.repository
+
+class MemoRepository {
+}
