+++ meme_android/app/src/main/java/com/likewhile/meme/ui/view/widget/MemoWidgetService.kt
@@ -1,7 +1,8 @@
-package com.likewhile.meme
+package com.likewhile.meme.ui.view.widget
 
 import android.content.Intent
 import android.widget.RemoteViewsService
+import com.likewhile.meme.ui.view.widget.MemoWidgetFactory
 
 class MemoWidgetService : RemoteViewsService() {
     override fun onGetViewFactory(intent: Intent?): RemoteViewsFactory {
