+++ meme_android/app/src/main/res/layout/activity_memo_edit.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MemoEditActivity">
+    tools:context=".ui.view.MemoEditActivity">
 
     <ScrollView
         android:id="@+id/scrollView2"
