+++ meme_android/app/src/main/res/layout/item_memo.xml
@@ -4,8 +4,8 @@
 
     <data>
         <import type="android.view.View" />
-        <import type="com.likewhile.meme.MemoItem"/>
-        <variable name="memo" type="MemoItem"/>
+        <import type="com.likewhile.meme.data.model.MemoItem"/>
+        <variable name="memo" type="com.likewhile.meme.data.model.MemoItem"/>
     </data>
 
     <androidx.cardview.widget.CardView
